+++ src/runtime/mod.rs
@@ -1,0 +1,6 @@
+// I know virtual_trackpad::start_handler
+// and VirtualTrackpad::clone are used
+// during initialization, but the rest
+// here is used in runtime only.
+pub mod event_handler;
+pub mod virtual_trackpad;